pyproject.toml
[project]
name = "interp-arena"
version = "0.1.0"
description = "Interpretability Arena: Red vs Blue — mechanistic adversarial LLM environment (OpenEnv)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-3-Clause" }

dependencies = [
    # OpenEnv
    "openenv-core>=0.2.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # Mechanistic interpretability
    "transformer-lens>=2.0.0",
    # ML
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    "datasets>=2.18.0",
    # Logging & config
    "wandb>=0.16.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    # RL / PEFT
    "trl>=0.9.0",
    "peft>=0.11.0",
    "bitsandbytes>=0.43.0",
    # Utilities
    "rich>=13.7.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "httpx>=0.27.0",
]
gpu = [
    # Run: pip install unsloth  (requires CUDA — install separately on GPU nodes)
    "trl>=0.9.0",
    "peft>=0.11.0",
    "bitsandbytes>=0.43.0",
]

[project.scripts]
train      = "scripts.train:main"
train-grpo = "scripts.train_grpo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["interp_arena"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -v"
asyncio_mode = "auto"
